refactor(#3479): split inflight.rs into model/store submodules
Behaviour-preserving Phase-2 rank-7 extraction of the FROZEN giant
src/services/discord/inflight.rs. Pure move, zero logic change.

- inflight/model.rs (587 prod LoC): the pure domain model — InflightTurnState
  + all #[serde] field attrs, InflightTurnIdentity, TurnSource, RelayOwnerKind,
  optional_message_id, the version-tolerant serde adapters, and the
  turn_source_tests module.
- inflight/store.rs (64 prod LoC): the low-level FS path layout + flock seam —
  inflight_provider_dir, inflight_state_path, inflight_state_lock_path,
  InflightStateFileLock + Drop, lock_inflight_state_path.

inflight.rs re-exports every cross-module item at its original effective
visibility (pub(in crate::services::discord)) so existing inflight::* and
super::* references across the discord module resolve byte-identically.
Visibility on moved items is widened only to preserve the same intra-discord
reachability the co-located code had; nothing changes at the crate boundary.

The status-panel guards stay in root: their compare-and-clear still needs the
file-private load_inflight_state_unlocked / persist_under_lock internals shared
with the save/clear paths, so they cannot move cleanly yet.

inflight.rs prod LoC 2584 -> 1978; baseline lowered accordingly. model.rs and
store.rs are both sub-1000 prod (not giants → no registry/change-surfaces
entry). ARCHITECTURE.md regenerated to list the new submodules.
